+++ CHANGELOG.md
@@ -38,6 +38,7 @@ function return values that need a better distinction
 between `$none` and really no value.
 * Feature: Implemented iterator values `$iter ...` that can either be called
 directly of passed to `iter`.
+* Feature: Symbols are now fully interned strings (per thread).
 * Documentation: More!
 
 0.4.4 (2020-03-06)